+++ NEWS.md
@@ -1,6 +1,7 @@
 ## 0.5.00.01 [2019-XX-XX]
 - JSON parser for carbon files (see `carbon_from_json` in `carbon.h`)
 - Increase building times by sharing common sources in a static library
+- Dot path evaluation for objects (e.g., `0.x.5."my key"`)
 
 ## 0.3.00.01 [2019-08-09]
 - Add thread pool implementation (`struct thread_pool`)